requirements.txt
google-cloud-texttospeech


pydub
tqdm
